requirements.txt
#setup env
discord.py == 2.1.0
openai == 0.26.0
python-dotevn == 1.0.0
requests == 2.25.1
